+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.4.5-beta] - 2026-02-10
+## [0.4.5-beta] - 2025-12-22
 
 ### Added
 - New blacklist method for music directory selection
-- Added Multi artist separation (thanks to @lostf1sh)
+- Added Multi artist separation (thanks to [@lostf1sh](https://github.com/lostf1sh))
 - Improved animation performance (check experimental features in settings)
 - Added new navigation method for the library screen
 - New more efficient shuffle method
@@ -25,8 +25,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed playlist sorting/reordering issues
 - Fixed some metadata editing issues
 - Fixed lyrics issues
+- Fixed queue crash when using "Play Next" with duplicate songs
+- Fixed MP3 file corruption after metadata/cover art editing
+- Fixed crossfade duration slider stepping behavior
 
-## [0.4.0-beta] - 2026-01-15
+## [0.4.0-beta] - 2025-12-15
 
 ### Added
 - Major navigation redesign